according to mvc model
